fix(): separate folder for archive and use snapshot for tests
